+++ macros/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump MSRV to 1.71.1.
 - All macros are now always available and all Cargo features have been removed.
+- The `bounds` options of derive macros has been removed.
 
 ## [0.8.0] - 2024-09-22